Refactored: moved Drive dialog related code to drive-dlg.cpp
