+++ src/CMakeLists.txt
@@ -140,6 +140,7 @@ SET(wcm_COMMON_SOURCES
 	folder-shortcuts.cpp
 	folder-history.cpp
 	view-history.cpp
+	drive-dlg.cpp
 	nceditline.cpp
 	smblogon.cpp
 	strconfig.cpp
@@ -223,6 +224,7 @@ SET(wcm_HEADERS
 	folder-shortcuts.h
 	folder-history.h
 	view-history.h
+	drive-dlg.h
 	nceditline.h
 	smblogon.h
 	strconfig.h
