+++ src/drive-dlg.h
@@ -1,0 +1,12 @@
+/*
+ * Part of WCM Commander
+ * https://github.com/corporateshark/WCMCommander
+ * walcommander@linderdaum.com
+ */
+
+#pragma once
+
+
+class PanelWin;
+
+void SelectDriveDlg( PanelWin* p, PanelWin* OtherPanel );
