+++ wcm_2013.vcxproj
@@ -270,6 +270,7 @@
     <ClCompile Include="src\urlparser\LUrlParser.cpp" />
     <ClCompile Include="src\folder-history.cpp" />
     <ClCompile Include="src\view-history.cpp" />
+    <ClCompile Include="src\drive-dlg.cpp" />
     <ClCompile Include="src\path-list.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -346,6 +347,7 @@
     <ClInclude Include="src/urlparser/LUrlParser.h" />
     <ClInclude Include="src/folder-history.h" />
     <ClInclude Include="src/view-history.h" />
+    <ClInclude Include="src/drive-dlg.h" />
     <ClInclude Include="src/path-list.h" />
   </ItemGroup>
   <ItemGroup>
